refactors test configuration
